Move GetHederaAccountId
